# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bike Trip Journal API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/trips/{slug}": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get a trip's metadata and its bikes",
        "description": "**Context.** A trip from before accounts was shared through two links, a rider\nlink and a viewer link. Since decision-log Entry 29 either link only *locates*\nthe trip; what the caller may do is decided by their membership. This is the\nendpoint both links open first. It answers three questions in one round trip \u2014\nwhich trip is this, which bikes are on it, and should the UI offer writes \u2014\nbecause the app is used on the Stuart Hwy where a second request is a second\nchance to be offline. Tasks `t-trip-metadata-endpoint`, `t-am-write-gate-legacy`.\n\n**How it works.** `{slug}` is resolved by the shared `require_trip_access`\ndependency, which accepts **either** slug and raises a 404 if neither matches.\nThe session is optional here and never produces a 401. The dependency derives\n`access`: `rider` iff the session user has an active membership on this trip\n(read fresh on every request), `viewer` for everyone else \u2014 anonymous, a\nsigned-in non-member, a pending requester or a revoked member. Which slug was\nfollowed plays no part. `viewer.role` is derived alongside it (`anonymous`,\n`none`, `pending`, `rider` or `leader`), and `access` is `rider` exactly when\nthat role is `rider` or `leader`. `visibility`, `publicDelayHours`, `riderCount`\nand `lastPublicStopAt` are filled as on `GET /api/v2/trips/{tripId}`; this read\nitself stays full and undelayed for either slug. The trip's bikes are fetched by `data/repositories/bikes.list_by_trip`,\nfiltered to this trip; a trip with no bikes returns `\"bikes\": []`, which is a\nnormal trip and not a 404. Neither slug is in the response: `TripOut` has no\nslug field, and the slug is the credential.\n\n`access` is a **UI hint**, deprecated, and not the enforcement point. It tells\nthe frontend whether to render Add stop / upload photo / edit bikes. The write\nendpoints check the session and membership themselves regardless of what the UI\nchose to show.\n\n**Related APIs.** `GET /api/trips/{slug}/stops` for the stops this header sits\nabove, `GET /api/trips/{slug}/map` for the same trip as GeoJSON, and\n`POST /api/trips/{slug}/bikes` / `PATCH /api/trips/{slug}/bikes/{id}` for the\nmember-only writes behind the `bikes` list returned here.",
        "operationId": "get_trip_api_trips__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The trip's rider or viewer slug \u2014 the link the trip was shared with. Either one locates the trip; neither grants anything. Legacy reads accept both and give the full, undelayed trip.",
              "title": "Slug"
            },
            "description": "The trip's rider or viewer slug \u2014 the link the trip was shared with. Either one locates the trip; neither grants anything. Legacy reads accept both and give the full, undelayed trip."
          }
        ],
        "responses": {
          "200": {
            "description": "The trip, the bikes on it, and whether the caller is an active member.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "404": {
            "description": "No trip has this slug. Deliberately the same answer for a mistyped link, a revoked one and a guess \u2014 see `docs/api-contract.md`, 'Access control: 403 and 404 are different answers'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "A path parameter failed validation. This operation takes no request body, so the path is the only thing that can fail; its path parameters are plain strings, so in practice it does not return this today.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "The `public-read` limit: 120 requests a minute per client address. Retry after `Retry-After` seconds.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait before retrying; always at least 1.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/trips/{tripId}": {
      "patch": {
        "tags": [
          "trips"
        ],
        "summary": "Update a trip's settings",
        "description": "**Context.** A leader's trip settings: its name, whether it is public, and how\nlong stops stay hidden from the public (decision-log Entry 29).\n\n**How it works.** `limit_writes` runs first, then `require_trip_leader`: a valid\nsession (401, for every trip id alike), then the trip (404 if it doesn't exist,\nor if it is private and the caller has no membership row on it -- the same 404\nthe v2 reads give), then an active **leader** membership (403). Only the fields\npresent in the body change; omitting a field is how \"no change\" is spelled, an\nexplicit `null` is a 422, and an empty body is a 200 with nothing changed.\nThere is no server-side publish confirmation. Last write wins.\n\n**Related APIs.** `GET /api/v2/trips/{tripId}` reads the settings back;\n`POST /api/v2/trips` creates the trip.",
        "operationId": "patch_trip_api_v2_trips__tripId__patch",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The trip's id. Not a secret: it only names the trip. Leader actions need a signed-in account with an active leader membership on it. No session is a 401 for every trip id, existing or not; a private trip the caller has no membership row on is a 404 identical to a trip id that doesn't exist; anyone else who is not an active leader -- a non-member of a public trip, a revoked member, an active rider -- is a 403.",
              "title": "Tripid"
            },
            "description": "The trip's id. Not a secret: it only names the trip. Leader actions need a signed-in account with an active leader membership on it. No session is a 401 for every trip id, existing or not; a private trip the caller has no membership row on is a 404 identical to a trip id that doesn't exist; anyone else who is not an active leader -- a non-member of a public trip, a revoked member, an active rider -- is a 403."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The trip after applying the patch.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "401": {
            "description": "No valid session: none sent, expired, signed out or revoked, or the account is disabled. Checked **before** the trip is looked up, so the answer is the same for a public trip, a private trip and a trip id that doesn't exist. If a session cookie was sent it is cleared (`Max-Age=0`). Nothing was changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, and the trip is visible to the caller (public, or they have a membership row on it), but they are not an active leader: \"You're not a rider on this trip.\" for a non-member of a public trip, \"You're no longer a rider on this trip.\" for a revoked member, \"You're not a leader on this trip.\" for an active rider. Nothing was changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No trip has this id, **or** the trip is private and the caller has no membership row on it. The two are byte-identical, as on the v2 reads, so a trip id can't be used to learn whether a private trip exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "The body failed validation: an explicit `null`, a `name` empty or over 100 characters after trimming, a `visibility` other than 'public'/'private', or a `publicDelayHours` outside 0-168.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "The `writes` limit: 600 requests an hour per account, or per client address for a request with no session. Checked before the session, the trip and the membership, so nothing was changed. Retry after `Retry-After` seconds.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait before retrying; always at least 1.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get a trip",
        "description": "**Context.** A trip's page, located by its id rather than a slug (decision-log\nEntry 29). Anyone can open a public trip; a private one is visible to its\nactive members only.\n\n**How it works.** `{tripId}` is resolved by `require_trip_reader`: the trip must exist and be\npublic, or the caller must be an active member (rider or leader); anything else\nis a 404 identical to a trip id that doesn't exist. The session is optional and\nnever produces a 401. Active members see everything with no delay. Anyone else\nsees only stops with `arrivedAt <= now() - publicDelayHours`, on the server\nclock at request time. `viewer.role` is computed from the session on every\nrequest: `anonymous`, `none`, `pending`, `rider` or `leader`. `access` is\ndeprecated and is `rider` exactly when the role is `rider` or `leader`.\n`riderCount` and `lastPublicStopAt` are the same for every caller, members\nincluded; `lastPublicStopAt` is the latest `arrivedAt` among stops visible to\nthe public. Bikes are not delayed. No slug, username or user id is in the\nresponse.\n\n**Related APIs.** `GET /api/v2/trips/{tripId}/stops`, `/map`, `/bikes` and\n`/stops/{stopId}/photos` for the rest of the trip; `GET /api/v2/trips` for the\npublic list.",
        "operationId": "get_trip_api_v2_trips__tripId__get",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The trip's id. Not a secret: it only names the trip. A private trip the caller may not see answers exactly as a trip id that doesn't exist does.",
              "title": "Tripid"
            },
            "description": "The trip's id. Not a secret: it only names the trip. A private trip the caller may not see answers exactly as a trip id that doesn't exist does."
          }
        ],
        "responses": {
          "200": {
            "description": "The trip, its bikes, and what the caller is to it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "404": {
            "description": "No trip has this id, **or** the trip is private and the caller is not an active member. The two are byte-identical -- status, body and headers -- so a trip id can't be used to learn whether a private trip exists. Never a 401: a read needs no session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "A path parameter failed validation. This operation takes no request body, so the path is the only thing that can fail; its path parameters are plain strings, so in practice it does not return this today.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "The `public-read` limit: 120 requests a minute per client address. Retry after `Retry-After` seconds.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait before retrying; always at least 1.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Access": {
        "type": "string",
        "enum": [
          "rider",
          "viewer"
        ],
        "title": "Access"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable \u2014 safe to show a rider directly."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "TripOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "The trip's id. Stable, and safe to show or log \u2014 it is not a credential and cannot be used to reach the trip. Neither slug appears anywhere in this response."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The trip's display name, as the rider titled it."
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "title": "Startdate",
            "description": "The day the trip starts, as an ISO 8601 date (YYYY-MM-DD). A date with no time and no timezone: it is the calendar day the journal opens on, not an instant, so it reads the same wherever the viewer is."
          },
          "bikes": {
            "items": {
              "$ref": "#/components/schemas/BikeOut"
            },
            "type": "array",
            "title": "Bikes",
            "description": "The bikes on **this** trip, embedded so the app's first load is one request rather than two. Empty list when no bikes have been added yet \u2014 that is a normal trip, not an error. Order is not part of the contract; match bikes by `id`."
          },
          "access": {
            "$ref": "#/components/schemas/Access",
            "description": "Deprecated; kept for clients from before accounts. 'rider' when the caller is signed in with an active membership on this trip (rider or leader), 'viewer' otherwise \u2014 anonymous, a signed-in non-member, a pending requester or a revoked member. Which slug was used plays no part. A UI hint for whether to show write UI (Add stop, upload photo, edit bikes); the server enforces access on every write regardless."
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility",
            "description": "Who can read the trip: 'public' (anyone, with the public delay applied for non-members) or 'private' (members only; anyone else gets the same 404 as for a trip that doesn't exist)."
          },
          "publicDelayHours": {
            "type": "integer",
            "title": "Publicdelayhours",
            "description": "How many hours a stop stays hidden from non-members after its `arrivedAt`, from 0 to 168. Members see every stop at once, whatever this says."
          },
          "riderCount": {
            "type": "integer",
            "title": "Ridercount",
            "description": "How many people are on the trip: its active members, leaders included. A count only; who they are is never in this response."
          },
          "lastPublicStopAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastpublicstopat",
            "description": "The latest `arrivedAt` among the stops visible to the public, i.e. with the public delay applied, as a timezone-aware ISO 8601 instant. The same value for every caller, members included. Null when no stop is visible to the public yet."
          },
          "viewer": {
            "$ref": "#/components/schemas/ViewerOut",
            "description": "What the caller is to this trip, computed from the session on every request. Replaces `access`. A UI hint only: the server enforces access on every write regardless."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "startDate",
          "bikes",
          "access",
          "visibility",
          "publicDelayHours",
          "riderCount",
          "lastPublicStopAt",
          "viewer"
        ],
        "title": "TripOut",
        "description": "One trip: `GET /api/trips/{slug}` (either slug) and `GET /api/v2/trips/{tripId}`.\n\nCarries no slug, no username and no user id -- a non-member can receive it."
      },
...
      "TripPatch": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "The trip's display name. Trimmed, then 1-100 characters, or the request is rejected with 422 / VALIDATION_ERROR. Omit the field to leave the stored value alone. Explicit null is rejected with 422 / VALIDATION_ERROR: omitting is how 'no change' is spelled."
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility",
            "description": "Who can read the trip: 'public' or 'private'. There is no server-side publish confirmation; the app warns before publishing. Omit the field to leave the stored value alone. Explicit null is rejected with 422 / VALIDATION_ERROR: omitting is how 'no change' is spelled."
          },
          "publicDelayHours": {
            "type": "integer",
            "maximum": 168.0,
            "minimum": 0.0,
            "title": "Publicdelayhours",
            "description": "How many hours a stop stays hidden from non-members after its `arrivedAt`, from 0 to 168 inclusive; anything else is rejected with 422 / VALIDATION_ERROR. 0 shows stops to the public as soon as they are sent. Omit the field to leave the stored value alone. Explicit null is rejected with 422 / VALIDATION_ERROR: omitting is how 'no change' is spelled."
          }
        },
        "type": "object",
        "title": "TripPatch",
        "description": "PATCH /api/v2/trips/{tripId} body. Leaders only. Every field may be omitted -- only\nfields present in the body change, and an empty body changes nothing -- but none may\nbe null."
      },
...
      "Visibility": {
        "type": "string",
        "enum": [
          "public",
          "private"
        ],
        "title": "Visibility",
        "description": "Who can read a trip. Mirrors `trips_visibility_check` in migration 0003."
      }
    }
  }
}